Bug fixes
Edited the seeders to have one active sprint (at time of commit).
Bug fixes when time estimates can be updated, when the input field shows up or when it doesn't.
Bug fixes when stories can be assigned to a sprint and when the checkboxes show up or when they're just disabled.